requirements.txt
pyrebase
python-docx
python-dotenv
google.ai.generativelanguage_v1beta.types
google.generativeai
flask
flask-cors
gunicorn